260220-tts-py/requirements.txt
pywebview
requests
beautifulsoup4
nltk
gTTS
pygame
